Add files via upload
I am not done with these files I have created a basic outline (just copied from my previous files but I have replaced many of the variable names in the form. I am created inputs and buttons right now, But I am out of time right now (2:20) will work on this later. I have not changed variable names for the top file, We need to create a navigation file which will not be hard. Just copy the one from our current websites and chage the variable names to the file names that we have now. I will work on this later. Thanks guys.